refactor: move textConnection to latex directory
