+++ coauthor-for-vs-code/src/latex/textConnection.ts
